[feat/#123] CourseRecommendGrid 파일 분리
