+++ Solply/Solply/Presentation/Course/Component/CourseRecommendGrid.swift
@@ -1,0 +1,40 @@
+//
+//  CourseRecommendGrid.swift
+//  Solply
+//
+//  Created by seozero on 7/15/25.
+//
+
+import SwiftUI
+
+struct CourseRecommendGrid: View {
+    
+    // MARK: - Properties
+    
+    private let columns = [
+        GridItem(.fixed(165.adjustedWidth), spacing: 12.5.adjustedWidth),
+        GridItem(.fixed(165.adjustedWidth))
+    ]
+    private let placeCategory: PlaceCategoryType = .all
+    
+    // MARK: - Body
+    
+    var body: some View {
+        VStack(alignment: .leading, spacing: 16.adjustedHeight) {
+            LazyVGrid(columns: columns, spacing: 13.adjustedHeight) {
+                ForEach(0..<10) { index in
+                    CourseCard(
+                        isSaved: true,
+                        title: "오감으로 수집하는 하루",
+                        courseCategory: [.book, .cafe],
+                        isSelected: true
+                    ) {
+                        print("ㅇㅇ")
+                    }
+                    .frame(width: 165.adjustedWidth, height: 165.adjustedHeight)
+                }
+            }
+        }
+        .padding(.horizontal, 20.adjustedWidth)
+    }
+}
